# docs/api-contract.openapi.yaml
openapi: "3.1.0"
info:
  title: "Club Informatique de l’IST : contrat d’API"
  version: "1.0.0"
  description: |
    Contrat de référence entre le frontend et le backend.

    Conventions : dates en ISO 8601 UTC ; pagination uniforme (« content », « page », « size », « totalElements », « totalPages ») ;
    erreurs au format RFC 9457 avec un code stable et des messages en français ; identifiants numériques stables.
    Le jeton d’accès est transmis par l’en-tête « Authorization » ; le jeton de rafraîchissement par un cookie HttpOnly.
    L’extension « x-roles » liste les rôles autorisés ; un Formateur et un Responsable du Club héritent des droits du Membre, un Super Admin de ceux de l’Administrateur.
    Les droits sont appliqués par le serveur, par rôle et par propriété de la ressource.
servers:
  - url: /api/v1
...
security:
  - jetonAcces: []
paths:
...
  "/bureau":
    get:
      tags: [Contenus publics]
      summary: Composition du bureau, par ordre d’affichage.
      operationId: getBureau
      security: []
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembreBureau"
    post:
      tags: [Gestion du club]
      summary: Ajouter un membre du bureau.
      operationId: postBureau
      x-roles: [RESPONSABLE_CLUB]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/MembreBureauSaisie"
      responses:
        "201":
          description: Ressource créée.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/MembreBureau"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
...
  "/gestion/indicateurs":
    get:
      tags: [Gestion du club]
      summary: Effectif actif et fréquentation mensuelle.
      operationId: getGestionIndicateurs
      x-roles: [RESPONSABLE_CLUB]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Indicateurs"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
...
  "/bureau/{id}":
    put:
      tags: [Gestion du club]
      summary: Modifier un membre du bureau.
      operationId: putBureauId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/MembreBureauSaisie"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/MembreBureau"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
    delete:
      tags: [Gestion du club]
      summary: Retirer un membre du bureau.
      operationId: deleteBureauId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Opération réalisée, sans contenu.
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
...
  "/admin/statistiques":
    get:
      tags: [Administration]
      summary: Totaux et répartitions.
      operationId: getAdminStatistiques
      x-roles: [ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Statistiques"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/security/alerts":
    get:
      tags: [Administration]
      summary: "Comptes signalés : verrouillages, suspensions, échecs de connexion répétés depuis 24 heures."
      operationId: getAdminSecurityAlerts
      x-roles: [ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alerte"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/security/audit-logs":
    get:
      tags: [Administration]
      summary: Journal d’audit.
      operationId: getAdminSecurityAuditLogs
      x-roles: [ADMIN]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: utilisateur
          in: query
          schema:
            type: string
          description: Adresse du compte.
        - name: statut
          in: query
          schema:
            type: string
            enum: [SUCCES, ECHEC]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageEntreeJournal"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/system/config":
    get:
      tags: [Système et conformité]
      summary: "Réglages persistants et appliqués : verrouillage des comptes, ouverture des inscriptions, mode maintenance (503 MAINTENANCE hors administration)."
      operationId: getAdminSystemConfig
      x-roles: [SUPER_ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Configuration"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
    put:
      tags: [Système et conformité]
      summary: Modifier les réglages.
      operationId: putAdminSystemConfig
      x-roles: [SUPER_ADMIN]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Configuration"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Configuration"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/system/sauvegardes":
    get:
      tags: [Système et conformité]
      summary: Dernières sauvegardes enregistrées par la tâche planifiée (décision D-09).
      operationId: getAdminSystemSauvegardes
      x-roles: [SUPER_ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sauvegarde"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/dsi/conformite":
    get:
      tags: [Système et conformité]
      summary: Contrôles de conformité calculés.
      operationId: getDsiConformite
      x-roles: [DSI]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Conformite"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/dsi/conformite/logs":
    get:
      tags: [Système et conformité]
      summary: Journal d’audit en consultation.
      operationId: getDsiConformiteLogs
      x-roles: [DSI]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageEntreeJournal"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
...
components:
  securitySchemes:
    jetonAcces:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Numéro de page, à partir de zéro.
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
      description: Taille de page.
    sort:
      name: sort
      in: query
      schema:
        type: string
      description: "Tri : « champ,asc » ou « champ,desc »."
...
  responses:
    Erreur400:
      description: "Requête invalide : erreurs de champ dans « errors »."
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur401:
      description: Session absente ou expirée, ou identifiants refusés.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur403:
      description: Droits insuffisants (rôle ou propriété de la ressource).
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur404:
      description: Ressource introuvable.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
...
  schemas:
    Probleme:
      type: object
      description: Erreur au format RFC 9457. Le champ « code » est stable ; « detail » est rédigé en français.
      required: [title, status, code]
      properties:
        type:
          type: [string, "null"]
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int64
        detail:
          type: [string, "null"]
        instance:
          type: [string, "null"]
        code:
          type: string
        errors:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/ErreurChamp"
    ErreurChamp:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    MembreBureau:
      type: object
      description: Aucune photo.
      required: [id, nom, prenom, fonction, ordre]
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        prenom:
          type: string
        fonction:
          type: string
        filiere:
          type: [string, "null"]
        ordre:
          type: integer
          format: int64
    MembreBureauSaisie:
      type: object
      required: [nom, prenom, fonction, ordre]
      properties:
        nom:
          type: string
        prenom:
          type: string
        fonction:
          type: string
        filiere:
          type: [string, "null"]
        ordre:
          type: integer
          format: int64
...
    Indicateurs:
      type: object
      description: Calculés en base, hors comptes de test. « frequentation » couvre les six derniers mois.
      required: [membresActifs, frequentation]
      properties:
        membresActifs:
          type: integer
          format: int64
        frequentation:
          type: array
          items:
            $ref: "#/components/schemas/PointFrequentation"
    PointFrequentation:
      type: object
      required: [mois, inscriptions]
      properties:
        mois:
          type: string
        inscriptions:
          type: integer
          format: int64
...
    Statistiques:
      type: object
      description: Agrégations calculées en base, hors comptes de test et éléments supprimés. « totalMembres » et « membresActifs » comptent tous les comptes réels, quel que soit leur rôle ; les comptes de test et ce qu’ils ont créé sont exclus de tous les totaux.
      required: [totalMembres, membresActifs, totalEvenements, totalFormations, totalProjets, totalRessources, totalMessagesNonTraites, repartitionMembresParRole, repartitionProjetsParStatut]
      properties:
        totalMembres:
          type: integer
          format: int64
        membresActifs:
          type: integer
          format: int64
        totalEvenements:
          type: integer
          format: int64
        totalFormations:
          type: integer
          format: int64
        totalProjets:
          type: integer
          format: int64
        totalRessources:
          type: integer
          format: int64
        totalMessagesNonTraites:
          type: integer
          format: int64
        repartitionMembresParRole:
          type: object
          additionalProperties:
            type: integer
            format: int64
        repartitionProjetsParStatut:
          type: object
          additionalProperties:
            type: integer
            format: int64
    Alerte:
      type: object
      required: [typeAlerte, description, gravite]
      properties:
        typeAlerte:
          type: string
        description:
          type: string
        utilisateurCible:
          type: [string, "null"]
          format: email
        utilisateurId:
          type: [integer, "null"]
          format: int64
        gravite:
          type: string
          enum: [FAIBLE, MOYENNE, CRITIQUE]
    EntreeJournal:
      type: object
      required: [id, action, dateAction, statut]
      properties:
        id:
          type: integer
          format: int64
        action:
          type: string
        description:
          type: [string, "null"]
        utilisateurEmail:
          type: [string, "null"]
          format: email
        ipAddress:
          type: [string, "null"]
        dateAction:
          type: string
          format: date-time
        statut:
          type: string
          enum: [SUCCES, ECHEC]
    Configuration:
      type: object
      required: [nomPlateforme, maintenanceMode, inscriptionsOuvertes, maxLoginAttempts, lockoutDurationMinutes]
      properties:
        nomPlateforme:
          type: string
        version:
          type: [string, "null"]
        maintenanceMode:
          type: boolean
        inscriptionsOuvertes:
          type: boolean
        maxLoginAttempts:
          type: integer
          format: int64
        lockoutDurationMinutes:
          type: integer
          format: int64
    Sauvegarde:
      type: object
      required: [id, date, statut]
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        tailleOctets:
          type: [integer, "null"]
          format: int64
        statut:
          type: string
          enum: [REUSSIE, ECHOUEE]
    Conformite:
      type: object
      required: [statut, versionBackend, versionJava, comptesActifs, tentativesEchouees, verifications]
      properties:
        statut:
          type: string
          enum: [CONFORME, A_EXAMINER]
        versionBackend:
          type: string
        versionJava:
          type: string
        comptesActifs:
          type: integer
          format: int64
        tentativesEchouees:
          type: integer
          format: int64
        verifications:
          type: array
          items:
            $ref: "#/components/schemas/Verification"
    Verification:
      type: object
      description: Chaque contrôle est calculé, jamais codé en dur.
      required: [code, libelle, conforme]
      properties:
        code:
          type: string
        libelle:
          type: string
        conforme:
          type: boolean
...
    PageEntreeJournal:
      type: object
      description: Page de résultats (pagination uniforme).
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/EntreeJournal"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer